requirement.txt
ultralytics==8.0.20
opencv-python==4.8.0.76
numpy==1.24.3
tensorflow==2.13.0
librosa==0.10.0.post2
pyserial==3.5